Fix postgres build, now using multi-stage docker build
This eliminates some of the excess logic required to build things.

+++ scripts/dev-extensions/build-docker-image.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
 
 docker build -t ${SIDECAR_EXTENSION_IMAGE_NAME}:${SIDECAR_EXTENSION_IMAGE_TAG} --rm -f Dockerfile       .
-docker build -t ${SIDECAR_SETUP_IMAGE_NAME}:${SIDECAR_SETUP_IMAGE_TAG}         --rm -f Dockerfile-setup .
+if test -n "${SIDECAR_SETUP_IMAGE_NAME:-}" ; then
+    docker build -t ${SIDECAR_SETUP_IMAGE_NAME}:${SIDECAR_SETUP_IMAGE_TAG}     --rm -f Dockerfile-setup .
+fi
